@@ -91,15 +86,28 @@
         ],
         "moduleNameMapper": {
             "\\.(css|less)$": "identity-obj-proxy",
-            "worker.js": "<rootDir>/__mocks__/workerMock.js"
+            "worker.js": "<rootDir>/__mocks__/workerMock.js",
+            "\\.svg$": "<rootDir>/__mocks__/SvgIcons.tsx"
         },
         "testResultsProcessor": "./node_modules/jest-junit-reporter"
     },
     "dependencies": {
-        "ramda": "^0.25.0",
+        "clean-webpack-plugin": "^3.0.0",
+        "fs-promise": "^2.0.3",
+        "localforage": "1.7.4",
+        "ramda": "^0.27.0",
         "react": "^16.13.1",
         "react-dom": "^16.13.1",
+        "react-redux": "^7.2.0",
+        "react-svg-loader": "^3.0.3",
+        "redux": "^4.0.5",
         "redux-devtools-extension": "^2.13.8",
-        "write-file": "1.0.0"
+        "redux-thunk": "^2.3.0",
+        "request-promise": "^4.2.5",
+        "typescript": "^3.9.5",
+        "uglifyjs-webpack-plugin": "^2.2.0",
+        "webpack-bundle-analyzer": "^3.8.0",
+        "write-file": "1.0.0",
+        "yargs": "^15.3.1"
     }
 }
