@@ -4,13 +4,9 @@ import Banner from '../../js/components/Banner';
 import { shallow } from 'enzyme';
 import { BANNER_TEXT } from '../../js/constants';
 
-jest.mock('../../icons/Offline.svg', () => {
-    return <div />;
-});
+jest.mock('../../icons/Offline.svg');
 
-jest.mock('../../icons/Online.svg', () => {
-    return <div />;
-});
+jest.mock('../../icons/Online.svg');
 
 describe('Banner component behaviour', () => {
 
