@@ -19,12 +19,11 @@ interface IBannerState {
  */
 export default class Banner extends React.Component<IBannerProps, IBannerState> {
 
-    state = {
-        hide: true,
-    };
-
-    dismiss = () => {
-        this.setState({ hide: true });
+    constructor(props) {
+        super(props);
+        this.state = {
+            hide: true,
+        };
     }
 
     componentDidUpdate(prevProps) {
@@ -36,18 +35,24 @@ export default class Banner extends React.Component<IBannerProps, IBannerState> 
         }
     }
 
+    dismiss = () => {
+        this.setState({ hide: true });
+    };
+
     render() {
         return (
-            <>{ this.state.hide || (this.props.hasNetwork && !this.props.isOffline) ?
-            null :
-                <div className={`offline-banner alert notification ${this.props.hasNetwork ? 'alert-online' : 'alert-offline'}`}>
-                    <span className="network-icon">{ this.props.hasNetwork ? <OnlineSvg height={30} /> : <OfflineSvg height={30} /> }</span>
-                    <span className="format-pre-line">{this.props.hasNetwork ? BANNER_TEXT.online : BANNER_TEXT.offline}</span>
-                    <button className="close" onClick={this.dismiss}>
-                        <span>{'\u00D7'}</span>
-                    </button>
-                </div>
-            }</>
+            <>
+                { this.state.hide || (this.props.hasNetwork && !this.props.isOffline) ?
+                    null : (
+                        <div className={`offline-banner alert notification ${this.props.hasNetwork ? 'alert-online' : 'alert-offline'}`}>
+                            <span className="network-icon">{ this.props.hasNetwork ? <OnlineSvg height={30} /> : <OfflineSvg height={30} /> }</span>
+                            <span className="format-pre-line">{this.props.hasNetwork ? BANNER_TEXT.online : BANNER_TEXT.offline}</span>
+                            <button className="close" onClick={this.dismiss}>
+                                <span>{'\u00D7'}</span>
+                            </button>
+                        </div>
+                    )}
+            </>
         );
     }
 }
