@@ -1,9 +1,14 @@
+export enum OfflineView {
+    cache = 0,
+    replay = 1,
+    noNetwork = 2,
+}
+
 export interface IOfflineState {
-    view?: number;
+    view?: OfflineView;
     status?: string;
     progress?: number;
     isDismissEnabled?: boolean;
-    hasInitialized?: boolean;
 }
 
 export interface IOfflineProps {
