@@ -1,7 +1,7 @@
 import { clone } from '../services/Utils';
-import { IState } from '../interfaces/IModels';
+import { IState, Id } from '../interfaces/IModels';
 
-declare var manywho: any;
+declare let manywho: any;
 
 let currentMapElementId = null;
 let id = null;
@@ -29,13 +29,13 @@ export const StateInit = (state: IState) => {
 /**
  * @param id
  */
-export const getStateValue = (id: any) => {
-    if (values[id.id]) {
-        const value = clone(values[id.id]);
+export const getStateValue = (valueId: Id) => {
+    if (values[valueId.id]) {
+        const value = clone(values[valueId.id]);
 
         // If required, get a property value from an object
-        if (id.typeElementPropertyId && value.objectData && value.objectData.length > 0) {
-            const property = value.objectData[0].properties.find(prop => prop.typeElementPropertyId === id.typeElementPropertyId);
+        if (valueId.typeElementPropertyId && value.objectData && value.objectData.length > 0) {
+            const property = value.objectData[0].properties.find((prop) => prop.typeElementPropertyId === valueId.typeElementPropertyId);
             if (property) {
                 value.contentValue = property.contentValue ? property.contentValue : null;
                 value.objectData = property.objectData ? property.objectData : null;
@@ -49,35 +49,36 @@ export const getStateValue = (id: any) => {
 };
 
 /**
- * @param id
+ * @param valueId
  * @param typeElementId
  * @param snapshot
  * @param value
  */
-export const setStateValue = (id: any, typeElementId: string, snapshot: any, value: any) => {
-    if (id.typeElementPropertyId) {
-        if (!values[id.id] || !values[id.id].objectData || values[id.id].objectData.length === 0) {
-            const typeElement = clone(snapshot.metadata.typeElements.find(type => type.id === typeElementId));
+export const setStateValue = (valueId: Id, typeElementId: string, snapshot: any, value: any) => {
+    if (valueId.typeElementPropertyId) {
+        if (!values[valueId.id] || !values[valueId.id].objectData || values[valueId.id].objectData.length === 0) {
+            const typeElement = clone(snapshot.metadata.typeElements.find((type) => type.id === typeElementId));
 
             typeElement.properties = typeElement.properties.map((property) => {
                 property.typeElementPropertyId = property.id;
                 delete property.id;
                 return property;
             });
+            typeElement.elementType = 'VARIABLE';
 
-            values[id.id] = {
+            values[valueId.id] = {
                 objectData: [typeElement],
             };
         }
 
-        const property = values[id.id].objectData[0].properties.find(prop => prop.typeElementPropertyId === id.typeElementPropertyId);
+        const property = values[valueId.id].objectData[0].properties.find((prop) => prop.typeElementPropertyId === valueId.typeElementPropertyId);
         if (property) {
             property.contentValue = value.contentValue;
             property.objectData = value.objectData;
         }
 
     } else {
-        values[id.id] = clone(value);
+        values[valueId.id] = clone(value);
     }
 };
 
@@ -89,8 +90,8 @@ export const setStateValue = (id: any, typeElementId: string, snapshot: any, val
  */
 export const StateUpdate = (inputs: any[], mapElement: any, snapshot: any) => {
     inputs.forEach((input) => {
-        const page = snapshot.metadata.pageElements.find(pageElement => pageElement.id === mapElement.pageElementId);
-        const component = page.pageComponents.find(pageComponent => pageComponent.id === input.pageComponentId);
+        const page = snapshot.metadata.pageElements.find((pageElement) => pageElement.id === mapElement.pageElementId);
+        const component = page.pageComponents.find((pageComponent) => pageComponent.id === input.pageComponentId);
 
         if (component.valueElementValueBindingReferenceId) {
             const value = snapshot.getValue(component.valueElementValueBindingReferenceId);
