@@ -56,65 +56,67 @@ const offlineState = (
         isReplaying: false,
         hasNetwork: true,
         isOffline: false,
-    },
-    action) => {
+    }, action,
+) => {
 
     switch (action.type) {
-    case 'HAS_NETWORK':
-        return {
-            ...state,
-            hasNetwork: true,
-        };
+        case 'HAS_NETWORK':
+            return {
+                ...state,
+                hasNetwork: true,
+            };
 
-    case 'HAS_NO_NETWORK':
-        return {
-            ...state,
-            hasNetwork: false,
-            isOffline: true,
-        };
+        case 'HAS_NO_NETWORK':
+            return {
+                ...state,
+                hasNetwork: false,
+                isOffline: true,
+            };
 
-    case 'IS_OFFLINE':
-        return {
-            ...state,
-            isOffline: true,
-            hasNetwork: action.payload,
-        };
+        case 'IS_OFFLINE':
+            return {
+                ...state,
+                isOffline: true,
+                hasNetwork: action.payload,
+            };
 
-    case 'IS_ONLINE':
-        return {
-            ...state,
-            isOffline: false,
-            hasNetwork: true,
-        };
+        case 'IS_ONLINE':
+            return {
+                ...state,
+                isOffline: false,
+                hasNetwork: true,
+            };
 
-    case 'IS_REPLAYING':
-        return {
-            ...state,
-            isReplaying: action.payload,
-        };
+        case 'IS_REPLAYING':
+            return {
+                ...state,
+                isReplaying: action.payload,
+            };
 
-    case 'CACHE_PROGRESS':
-        return {
-            ...state,
-            cachingProgress: action.payload,
-        };
+        case 'CACHE_PROGRESS':
+            return {
+                ...state,
+                cachingProgress: action.payload,
+            };
 
-    case 'FLOW_INFORMATION':
-        return {
-            ...state,
-            flowInformation: {
-                tenantId: action.payload.tenantId,
-                stateId: action.payload.stateId,
-                token: action.payload.token,
-            },
-        };
-    case 'POLLING_VALUES':
-        return {
-            ...state,
-            isPollingValues: action.payload,
-        };
-    default:
-        return state;
+        case 'FLOW_INFORMATION':
+            return {
+                ...state,
+                flowInformation: {
+                    tenantId: action.payload.tenantId,
+                    stateId: action.payload.stateId,
+                    token: action.payload.token,
+                },
+            };
+
+        case 'POLLING_VALUES':
+            return {
+                ...state,
+                isPollingValues: action.payload,
+            };
+
+        default:
+            return state;
     }
 };
 
