@@ -13,7 +13,7 @@ import { append, clone } from 'ramda';
  * cached data and setting to appropriate value
  */
 const loadData = (action: any, objectData: any, snapshot: any) => {
-    const filteredObjectData = ObjectData.filter(objectData, action.objectDataRequest.listFilter, action.objectDataRequest.typeElementId);
+    const filteredObjectData = ObjectData.filter(objectData, action.objectDataRequest.listFilter);
     const value = snapshot.getValue(action.valueElementToApplyId);
     setStateValue(action.valueElementToApplyId, value.typeElementId, snapshot, filteredObjectData);
 };
@@ -27,9 +27,9 @@ const loadData = (action: any, objectData: any, snapshot: any) => {
  */
 const saveData = (action: any, objectData: any, snapshot: any) => {
     const valueReferenceToSave = snapshot.getValue(action.valueElementToApplyId);
-    const typeElementId = valueReferenceToSave.typeElementId;
-    const type = typeElementId ? snapshot.metadata.typeElements.find(typeElement => typeElement.id === typeElementId) : null;
-    const valueElementToApplyId = action.valueElementToApplyId;
+    const { typeElementId } = valueReferenceToSave;
+    const type = typeElementId ? snapshot.metadata.typeElements.find((typeElement) => typeElement.id === typeElementId) : null;
+    const { valueElementToApplyId } = action;
 
     const valueToSave = getStateValue(valueElementToApplyId);
 
@@ -50,6 +50,8 @@ const saveData = (action: any, objectData: any, snapshot: any) => {
             if (obj.externalId && existingObj.objectData.externalId === obj.externalId) {
                 return existingObj;
             }
+
+            return null;
         });
 
         // Objectdata that has already been added but is now being modified
@@ -89,7 +91,7 @@ const saveData = (action: any, objectData: any, snapshot: any) => {
                 isSelected: false,
                 properties: clone(type.properties).map((property) => {
                     const newProp = obj.properties.filter(
-                        prop => prop.typeElementPropertyId === property.id,
+                        (prop) => prop.typeElementPropertyId === property.id,
                     );
                     if (newProp.length > 0) {
                         property.contentValue = newProp[0].contentValue ? newProp[0].contentValue : null;
@@ -126,20 +128,23 @@ const saveData = (action: any, objectData: any, snapshot: any) => {
 export default (action: any, flow: IFlow, snapshot: any) => {
     const objectData = getObjectData(
         action.objectDataRequest.objectDataType ?
-        action.objectDataRequest.objectDataType.typeElementId :
-        action.objectDataRequest.typeElementId,
+            action.objectDataRequest.objectDataType.typeElementId :
+            action.objectDataRequest.typeElementId,
     );
 
     switch (action.crudOperationType.toUpperCase()) {
-    case 'LOAD':
-        loadData(action, objectData, snapshot);
-        break;
-    case 'SAVE':
-        saveData(action, objectData, snapshot);
-        break;
-    case 'DELETE':
+        case 'LOAD':
+            loadData(action, objectData, snapshot);
+            break;
+        case 'SAVE':
+            saveData(action, objectData, snapshot);
+            break;
+        case 'DELETE':
         // No implementation for a delete as its potential very destructive
-        break;
+            break;
+        default:
+            // TODO - Exception ?
+            break;
     }
 
     return flow.state;
