@@ -1,4 +1,4 @@
-import { clone } from '../services/Utils';
+import { clone } from './Utils';
 import { getStateValue } from '../models/State';
 import Rules from './Rules';
 
@@ -13,7 +13,7 @@ const ObjectData = {
      * @param objectData
      * @param filter
      */
-    filter: (objectData: any, filter: any, typeElementId: string = null) => {
+    filter: (objectData: any, filter: any) => {
         if (!filter || !objectData) {
             return {
                 objectData,
@@ -21,7 +21,7 @@ const ObjectData = {
             };
         }
 
-        const clonedObjectData = clone(objectData.map((obj) => { return obj.objectData; }));
+        const clonedObjectData = clone(objectData.map((obj) => obj.objectData));
         let filteredObjectData = [];
 
         // Support for where filtering
@@ -30,7 +30,7 @@ const ObjectData = {
                 const comparer = filter.comparisonType === 'OR' ? 'some' : 'every';
 
                 return filter.where[comparer]((where) => {
-                    const property = item.properties.find(property => property.typeElementPropertyId === where.columnTypeElementPropertyId);
+                    const property = item.properties.find((prop) => prop.typeElementPropertyId === where.columnTypeElementPropertyId);
 
                     if (!property) {
                         return true;
@@ -57,21 +57,20 @@ const ObjectData = {
         }
 
         if (filter.search) {
-            filteredObjectData = clonedObjectData.filter((item) => {
-                return item.properties.filter((property) => {
-                    return property.contentValue && property.contentValue.toLowerCase().indexOf(filter.search.toLowerCase()) !== -1;
-                }).length > 0;
-            });
+            filteredObjectData = clonedObjectData.filter((item) => item.properties
+                .filter((property) => property.contentValue &&
+                    property.contentValue.toLowerCase().indexOf(filter.search.toLowerCase()) !== -1).length > 0);
         }
 
         filter.offset = filter.offset || 0;
-        const page = Math.ceil(filter.offset / filter.limit);
+        const limit:number = +filter.limit || 0;
+        const page = Math.ceil(filter.offset / limit);
 
-        const slicedObjectData = isNaN(page) ?
-        filteredObjectData : filteredObjectData.slice(page * filter.limit, (page * filter.limit) + filter.limit);
+        const slicedObjectData = page ? filteredObjectData.slice(page * filter.limit, (page * limit) + limit) :
+            filteredObjectData;
 
         return {
-            hasMoreResults: ((page * filter.limit) + filter.limit + 1) <= filteredObjectData.length,
+            hasMoreResults: ((page * limit) + limit + 1) <= filteredObjectData.length,
             objectData: slicedObjectData,
         };
     },
