@@ -1,4 +1,4 @@
-import { clone } from '../../services/Utils';
+import { clone } from '../Utils';
 
 const currentState: any = {};
 
@@ -6,16 +6,14 @@ const currentState: any = {};
  * @description returns an object of state values
  * that have been modified by macro methods
  */
-export const getMacroState = () => {
-    return clone(currentState);
-};
+export const getMacroState = () => clone(currentState);
 
 /**
  * @param state a modified copy of state values
  * @description mutates current stored state values
  * into state values passed in
  */
-export const setMacroState = (state: object) => {
+export const setMacroState = (state: any) => {
     for (const key of Object.keys(state)) {
         currentState[key] = state[key];
     }
