@@ -20,49 +20,42 @@ const PageRule = (
     // and is_empty criteria types
     switch (criteriaType) {
 
-    case CRITERIA.isEqual:
-        if (
-            String(triggerComponentContentValue).toUpperCase() ===
-            String(valueComparable.defaultContentValue).toUpperCase()
-        ) {
-            return true;
-        }
-        return false;
+        case CRITERIA.isEqual:
+            return (String(triggerComponentContentValue).toUpperCase() === String(valueComparable.defaultContentValue).toUpperCase());
 
-    case CRITERIA.isNotEqual:
-        if (
-            String(triggerComponentContentValue).toUpperCase() !==
-            String(valueComparable.defaultContentValue).toUpperCase()
-        ) {
-            return true;
-        }
-        return false;
+        case CRITERIA.isNotEqual:
+            return (String(triggerComponentContentValue).toUpperCase() !== String(valueComparable.defaultContentValue).toUpperCase());
 
-    case CRITERIA.isEmpty:
+        case CRITERIA.isEmpty:
 
-        // Is empty
-        if (String(valueComparable.defaultContentValue).toUpperCase() === 'TRUE') {
-            if (
-                (String(triggerComponentContentValue).toUpperCase() === 'NULL' ||
+            // Is empty
+            if (String(valueComparable.defaultContentValue).toUpperCase() === 'TRUE') {
+                if (
+                    (String(triggerComponentContentValue).toUpperCase() === 'NULL' ||
                 triggerComponentContentValue === '') &&
                 value.objectData === null
-            ) {
-                return true;
-            }
-            return false;
-        }
-
-        // Is not empty
-        if (String(valueComparable.defaultContentValue).toUpperCase() === 'FALSE') {
-            if (
-                (String(triggerComponentContentValue).toUpperCase() === 'NULL' ||
-                triggerComponentContentValue === '') &&
-                value.objectData === null
-            ) {
+                ) {
+                    return true;
+                }
                 return false;
             }
-            return true;
-        }
+
+            // Is not empty
+            if (String(valueComparable.defaultContentValue).toUpperCase() === 'FALSE') {
+                if (
+                    (String(triggerComponentContentValue).toUpperCase() === 'NULL' ||
+                triggerComponentContentValue === '') &&
+                value.objectData === null
+                ) {
+                    return false;
+                }
+                return true;
+            }
+            // TODO - What if valueComparable.defaultContentValue is neither TRUE or FALSE
+            return false;
+
+        default:
+            return false;
     }
 };
 
