@@ -1,4 +1,6 @@
-const enzyme = require("enzyme");
-const Adapter = require("enzyme-adapter-react-16");
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const enzyme = require('enzyme');
+const Adapter = require('enzyme-adapter-react-16');
 
 enzyme.configure({ adapter: new Adapter() });
